DOPSS-2 Profile Update fix

+++ Python_Django/DOPPS/blog/models.py
@@ -2,19 +2,23 @@ from django.db import models
 from django.contrib.auth.models import User
 from django.urls import reverse
 from django.utils.timezone import now
+from django.core.validators import MaxValueValidator, MinValueValidator
 
 # Create your models here.
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, blank=True, null=True)
     image = models.ImageField(upload_to="profile_pics", blank=True, null=True)
     bio = models.TextField(blank=True, null=True)
-    phone_no = models.IntegerField(blank=True, null=True)
+    phone_no = models.PositiveSmallIntegerField(blank=True, null=True)
     facebook = models.CharField(max_length=300, blank=True, null=True)
     instagram = models.CharField(max_length=300, blank=True, null=True)
     linkedin = models.CharField(max_length=300, blank=True, null=True)
     
     def __str__(self):
         return str(self.user)
+    
+    # def get_absolute_url(self):
+    #     return reverse('blogs')
  
 class BlogPost(models.Model):
     title=models.CharField(max_length=255)
